+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  *
+  * Do not remove `public/system` from the repository when it exists. Warn that maintenance pages may not work.
 
 ## v2.3.9 (2014-03-26)
 
@@ -42,7 +42,7 @@
 
 ## v2.3.2 (2013-10-09)
 
-  * Add hooks `before_deploy` and `after_deploy` that happen near the beginning (after release directory is created) and at the end of the deploy task. 
+  * Add hooks `before_deploy` and `after_deploy` that happen near the beginning (after release directory is created) and at the end of the deploy task.
   * Support option --clean which will force gems to be reinstalled (eg. when making a server architecture change that isn't already caught)
   * If dependncy manager options are set to blank, treat it like "detect".
   * Fix an issue with integrate action.